+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ### Changed
 
+- Bach database operations in mappings
+
 ### Fixed
 
 ## [17.0.22] - 2026-05-21